refactoring code, add status in Driver Licence, add full access for Admin, create CarDTO and change logic Car Service and Car Controller

+++ src/main/java/rentcar/rentcar/controller/FineController.java
@@ -17,7 +17,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import rentcar.rentcar.domain.Fine;
-import rentcar.rentcar.dto.FineDTO;
 import rentcar.rentcar.service.FineService;
 
 import javax.validation.Valid;
@@ -44,13 +43,17 @@ public class FineController {
 
     @GetMapping("/all")
     public ResponseEntity<ArrayList<Fine>> getAllFines() {
-        return new ResponseEntity<>(fineService.getAllFine(), HttpStatus.OK);
+        if(fineService.getAllFine() != null) {
+            return new ResponseEntity<>(fineService.getAllFine(), HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
     }
 
     @PostMapping("/create")
-    public ResponseEntity<HttpStatus> createFine(@RequestBody @Valid FineDTO fine, BindingResult bindingResult) {
-        if(bindingResult.hasErrors()) {
-            for(ObjectError o : bindingResult.getAllErrors()) {
+    public ResponseEntity<HttpStatus> createFine(@RequestBody @Valid Fine fine, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            for (ObjectError o : bindingResult.getAllErrors()) {
                 log.warn(o.getDefaultMessage());
             }
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
@@ -60,23 +63,21 @@ public class FineController {
     }
 
     @PutMapping("/update")
-    public ResponseEntity<HttpStatus> updateFine(@RequestBody @Valid FineDTO fine, BindingResult bindingResult) {
-        if(bindingResult.hasErrors()) {
-            for(ObjectError o : bindingResult.getAllErrors()) {
-                log.warn(o.getDefaultMessage());
-            }
+    public ResponseEntity<HttpStatus> updateFine(@RequestBody Fine fine) {
+        if (fineService.updateFine(fine)) {
+            fineService.updateFine(fine);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
-        fineService.updateFine(fine);
-        return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @DeleteMapping("delete/{id}")
     public ResponseEntity<HttpStatus> deleteFine(@PathVariable int id) {
-        try {
+        if (fineService.deleteFine(id)) {
             fineService.deleteFine(id);
             return new ResponseEntity<>(HttpStatus.OK);
-        } catch (IllegalArgumentException e) {
+        } else {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
     }
